chore: 重构demo
