+++ packages/maps/src/amap2/map.ts
@@ -20,6 +20,7 @@ import 'reflect-metadata';
 import type { IAMapInstance } from '../../typings/index';
 import AMapBaseService from '../utils/amap/AMapBaseService';
 import Viewport from './Viewport';
+// @ts-ignore
 import './logo.css';
 const DEFAULTMAPCENTER = [108.92361, 34.54083];
 
@@ -371,7 +372,7 @@ export default class AMapService extends AMapBaseService {
       // @ts-ignore
       // left, right, bottom, top
       // @ts-ignore 
-    } = this.map.customCoords?.getCameraParams();
+    } = this.map.customCoords?.getCameraParams() || {};
     // Tip: 统一触发地图变化事件
     this.emit('mapchange');
     // // @ts-ignore
